refactor to make it easier to make custom icons
