+++ react-app/src/icons/FromLink/index.ts
@@ -1,1 +1,2 @@
-export { default } from './FromLink';
+export { default } from './IconFromLink';
+export type { IconProps as IconFromLinkProps } from '@material-ui/core/Icon';
